+++ properties.py
@@ -17,8 +17,6 @@
 # ##### END GPL LICENSE BLOCK #####
 
 import bpy
-import os
-import re
 
 #
 # Custom Property Groups
@@ -84,10 +82,3 @@ class PreflightOptionsGroup(bpy.types.PropertyGroup):
         default="//preflight",
         maxlen=1024,
         subtype='DIR_PATH')
-
-
-def register():
-    bpy.utils.register_module(__name__)
-
-def unregister():
-    bpy.utils.unregister_module(__name__)
